+++ dart/simulation/experimental/compute/multibody_dynamics.cpp
@@ -437,8 +437,9 @@ DynamicsTree buildDynamicsTree(
     dynamics.jointEntity = linkComp.parentJoint;
     tree.jointOf[i] = linkComp.parentJoint;
 
-    const Eigen::Isometry3d childInParent
-        = jointMotionTransform(joint) * linkComp.transformFromParentJoint;
+    const Eigen::Isometry3d childInParent = linkComp.transformToParentJoint
+                                            * jointMotionTransform(joint)
+                                            * linkComp.transformFromParentJoint;
     dynamics.parentToChild = adjoint(childInParent.inverse());
     dynamics.childToParentForce = dynamics.parentToChild.transpose();
     dynamics.worldTransform